Add support for client cert auth when validating.
The following directives were added:
    - CASClientCert - the client cert in PEM or DER format
    - CASClientKey - the client key in PEM or DER format
    - CASClientCertType - PEM or DER - the cert/key type

Configuration:
    CASClientCertType "PEM"
    CASClientCert /path/to/client/cert.pem
    CASClientKey /path/to/client/key.pem

The paths must point to valid files on startup. No checking is currently done to
ensure they exist when validating. This check could probably be added at some
point.

A key password directive could be added, but it seems kind of silly to me. Up
for debate.